refactor(pilots): decompose PilotCreationWizard into step sub-components
- PilotCreationWizard.tsx: 1,181 → 403 lines (66% reduction)
- Created wizard/ subdirectory with 7 files:
  - WizardTypes.ts (188L) - types, constants, MODE_INFO, utilities
  - StepIndicator.tsx (84L) - step progress indicator
  - ModeSelectionStep.tsx (121L) - mode selection cards
  - IdentityStep.tsx (86L) - pilot identity form
  - SkillsStep.tsx (315L) - template/custom/random skills config
  - ReviewStep.tsx (119L) - review summary + statblock warning
  - index.ts (17L) - barrel exports
- Shell at 403 lines (state management + handlers + modal chrome)
- All 28 characterization tests pass
- Zero TypeScript errors, build successful
