+++ src/components/pilots/wizard/index.ts
@@ -1,0 +1,17 @@
+export { StepIndicator } from './StepIndicator';
+export { ModeSelectionStep } from './ModeSelectionStep';
+export { IdentityStep } from './IdentityStep';
+export { SkillsStep, SkillSlider, SkillPreview } from './SkillsStep';
+export { ReviewStep } from './ReviewStep';
+export type {
+  CreationMode,
+  WizardStep,
+  WizardState,
+  PilotCreationWizardProps,
+} from './WizardTypes';
+export {
+  INITIAL_STATE,
+  MODE_INFO,
+  generateRandomSkills,
+  calculatePointCost,
+} from './WizardTypes';
